Rename the module from CLPIR to ir
