+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "clp_ffi_py"
-version = "0.0.0"
+version = "0.0.1"
 license = { text = "Apache License 2.0" }
 authors = [
     { name="zhihao lin", email="lin.zhihao@yscope.com" },
